+++ src/data/navigation/header.yml
@@ -4,5 +4,7 @@ items:
     link: /data
   - title: Historical data
     link: /data/historical
+  - title: Why it matters
+    link: /why-it-matters
   - title: How you can help
     link: /help
